Extend implementation to long
- Mark As and Try methods for AggressiveInlining
- Use As() for casts

`int` in and out is now less than a nanosecond when using Try().

Using the cast operators is rougly 5 times slower. Couldn't find any [MethodImpl] options that made an impact. This is likely due to the fact the Value struct is copied.

As() is about twice the cost of Try(). The JIT apparently doesn't like to inline two levels deep.