Add support for the rest of the primitives
Decimal has to be boxed as it won't fit. Could do special handling if it could fit into a long, but that probably isn't worth the effort. If it can be shown to be common that they fit into a long/ulong then it could be investigated. Since we'd be flagging we could expand the unboxed range a bit by using multiple flags...